+++ src/table/Table.js
@@ -9,15 +9,13 @@ import {
   isTableEditable,
   isRowEditable,
   skipOverride,
-  shouldDisplay,
-  isTableFooterShown
+  shouldDisplay
 } from '../Utils'
 import * as R from 'ramda'
 import DetailLink from '../DetailLink'
 import { Link } from 'react-router-dom'
 import { getModel, getActions, getFieldConditions } from '../utils/schemaGetters'
-import { DeleteDetail } from '../delete/DeleteDetail'
-
+import { DeleteDetail, RemoveDetail } from '../delete/DeleteDetail'
 import {
   RowEditButton,
   EditSaveButton,
@@ -30,6 +28,7 @@ import {
 import getDisplayValue from '../utils/getDisplayValue'
 import { IndexPagination, DetailPagination } from '../Pagination'
 import { getFieldOrderAlternate } from './FiledOrderAlter'
+import { getType } from '../utils/getType'
 
 export const DetailViewButton = ({ modelName, id }) => (
   <Link to={`/${modelName}/${id}`} className="btn btn-sm btn-outline-primary">
@@ -46,6 +45,18 @@ export const DeleteButton = ({ modalId, onDeleteWarning, modelName, id }) => {
       onClick={() => onDeleteWarning({ modelName, id })}
     >
       Delete
+    </button>
+  )
+}
+
+export const RemoveButton = ({ modalId }) => {
+  return (
+    <button
+      className="btn btn-sm btn-outline-warning"
+      data-toggle="modal"
+      data-target={'#' + modalId}
+    >
+      Remove
     </button>
   )
 }
@@ -68,11 +79,16 @@ export const TableButtonGroup = ({
   parentFieldName,
   deletable,
   onDelete,
+  fromIndex,
   customProps
 }) => {
+  const parentFieldType = getType({ schema, modelName: parentModelName, fieldName: parentFieldName })
+  const m2m = parentFieldType === 'ManyToMany'
   const actions = getActions(schema, modelName)
-  const modalId = 'confirm-delete-' + modelName + parentFieldName + idx
+  const onRemove = R.path(['edit', 'onDetailTableRemoveSubmit'], actions)
+  const modalId = `confirm-${m2m ? 'remove' : 'delete'}-${modelName}-${parentFieldName}-${idx}`
   const id = node.id
+  const canRemove = !fromIndex && m2m && editable
   return (
     <React.Fragment>
       <div className="btn-group">
@@ -88,7 +104,16 @@ export const TableButtonGroup = ({
             }}
           />
         )}
-        {deletable && (
+        {canRemove && (
+          <RemoveButton
+            {...{
+              modalId,
+              modelName,
+              id
+            }}
+          />
+        )}
+        {deletable && !canRemove && (
           <DeleteButton
             {...{
               modalId,
@@ -99,7 +124,29 @@ export const TableButtonGroup = ({
           />
         )}
       </div>
-      {deletable && (
+      {canRemove && (
+        <RemoveDetail
+          {...{
+            schema,
+            id,
+            modalId,
+            modelName,
+            onRemove,
+            parentId,
+            parentModelName,
+            parentFieldName,
+            name: getDisplayValue({
+              schema,
+              modelName,
+              parentModelName,
+              node,
+              customProps
+            }),
+            customProps
+          }}
+        />
+      )}
+      {deletable && !canRemove && (
         <DeleteDetail
           {...{
             schema,
@@ -226,6 +273,7 @@ export const TableButtonCell = ({
   parentFieldName,
   onDelete,
   idx,
+  fromIndex,
   customProps
 }) => {
   return isEditing(editData, modelName, node.id) ? (
@@ -258,6 +306,7 @@ export const TableButtonCell = ({
         parentModelName,
         parentFieldName,
         onDelete,
+        fromIndex,
         customProps
       }}
     />
@@ -366,6 +415,7 @@ const TBody = ({
                       parentFieldName,
                       onDelete,
                       idx,
+                      fromIndex,
                       customProps
                     }}
                   />
@@ -454,9 +504,6 @@ export const Table = ({
     customProps
   })
 
-  const footerShown = fromIndex && isTableFooterShown({ schema, modelName, user })
-
-  // TODO: Add Footer with summation if the view has showFooter enabled for at least one item
   return (
     <React.Fragment>
       <table className="table table-striped table-bordered table-hover">
@@ -501,23 +548,18 @@ export const Table = ({
             customProps
           }}
         />
-        {footerShown && <Foot
+        <Foot
           {...{
             schema,
             modelName,
+            parentModelName,
+            parentFieldName,
             fieldOrder,
-            editable,
-            deletable,
-            detailField,
-            selectOptions,
-            data,
-            tableView,
+            summary,
             fromIndex,
             customProps,
-            user,
-            summary
           }}
-        />}
+        />
       </table>
       {fromIndex ? (
         <IndexPagination