+++ CHANGES.md
@@ -4,7 +4,11 @@
 
 ### Unreleased
 
-- Add Summary and Footer #141
+-   Add Summary and Footer #141
+-   Add 'go to' pagination tooltip #146
+-   Fix multi-select bug #157
+-   Add Removable feature #159
+-   Fix target name in 'Remove' modal #163
 
 ## Version 1.5.0